third_party/WebKit/LayoutTests/inspector/tracing/timeline-time-expected.txt

Test nesting of time/timeEnd records on Timeline


Running: testSimpleConsoleTime
----> Function Call

Running: testNestedConsoleTime
----> Function Call

Running: testUnbalancedConsoleTime
----> Function Call

Running: testConsoleTimeWithoutConsoleTimeEnd
----> Function Call
--------> Timestamp: Foo
--------> Timestamp: Bar
--------> Timestamp: Baz
